requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-socketio[client]>=5.11.0
requests>=2.31.0
aiohttp>=3.9.0
